+++ RELEASE_NOTES.md
@@ -2,7 +2,8 @@ Release Notes
 =============
 
 ## 1.6.33
-* Container Registry name validation
+* Container Registry: Adds name validation
+* DNS: Add support for private DNS zones and records
 * PostgreSQL: Added possibility to set vnet rules for PostgreSQL.
 
 ## 1.6.32